Refactor: move the Polygon test to its own file. Reuse methods between the two files via new util.py for tests
