+++ tests/integration/remote/test_mumbai.py
@@ -2,17 +2,15 @@
 # Copyright 2022 Ocean Protocol Foundation
 # SPDX-License-Identifier: Apache-2.0
 #
-import os
 import random
-import string
 import warnings
 
 import pytest
 import requests
 import time
 
+from .util import get_wallets, random_chars
 from ocean_lib.ocean.ocean import Ocean
-from ocean_lib.web3_internal.wallet import Wallet
 
 
 def test_nonocean_tx(tmp_path):
@@ -21,7 +19,7 @@ def test_nonocean_tx(tmp_path):
     # setup
     config = _remote_config_mumbai(tmp_path)
     ocean = Ocean(config)
-    (alice_wallet, bob_wallet) = _get_wallets(ocean)
+    (alice_wallet, bob_wallet) = get_wallets(ocean)
 
     # Get gas price (in Gwei) from Polygon gas station
     gas_station_url = "https://gasstation-mumbai.matic.today/v2"
@@ -63,7 +61,7 @@ def test_nonocean_tx(tmp_path):
 
 
 @pytest.mark.skip(reason="Don't skip, once fixed #943")
-def test_ocean_tx1(tmp_path):
+def test_ocean_tx__create_data_nft(tmp_path):
     """On Mumbai, do a simple Ocean tx: create_data_nft"""
 
     # setup
@@ -73,7 +71,7 @@ def test_ocean_tx1(tmp_path):
 
     # Alice publish data NFT
     # avoid "replacement transaction underpriced" error: make each tx diff't
-    symbol = _randomChars()
+    symbol = random_chars()
     try:  # it can get away with "insufficient funds" errors, but not others
         print("Call create_data_nft(), and wait for it to complete...")
         data_nft = ocean.create_data_nft(symbol, symbol, alice_wallet)
@@ -88,63 +86,6 @@ def test_ocean_tx1(tmp_path):
     print("Success")
 
 
-@pytest.mark.integration
-def test_ocean_tx2(tmp_path):
-    """On Mumbai, do the Ocean txs for create_url_asset(). Captures issue:https://github.com/oceanprotocol/ocean.py/issues/1007#issuecomment-1276286245"""
-
-    # setup
-    config = _remote_config_polygon(tmp_path)
-    ocean = Ocean(config)
-    (alice_wallet, _) = _get_wallets(ocean)
-
-    # Alice call create_url_asset
-    # avoid "replacement transaction underpriced" error: make each tx diff't
-    cand_chars = string.ascii_uppercase + string.digits
-    name = _randomChars()
-    url = "https://arweave.net/qctEbPb3CjvU8LmV3G_mynX74eCxo1domFQIlOBH1xU"
-    try:  # it can get away with "insufficient funds" errors, but not others
-        print("Call create_url_asset(), and wait for it to complete...")
-        asset = ocean.assets.create_url_asset(name, url, alice_wallet)
-
-    except ValueError as error:
-        if "insufficient funds" in str(error):
-            warnings.warn(UserWarning("Warning: Insufficient test MATIC"))
-            return
-        raise (error)
-
-    assert asset is not None
-    datatoken_address = asset.datatokens[0]["address"]
-    datatoken = ocean.get_datatoken(datatoken_address)
-    assert datatoken is not None
-    print("Success")
-
-
-def _randomChars() -> str:
-    cand_chars = string.ascii_uppercase + string.digits
-    return "".join(random.choices(cand_chars, k=8))
-
-
-def _get_wallets(ocean):
-    config, web3 = ocean.config_dict, ocean.web3
-
-    alice_private_key = os.getenv("REMOTE_TEST_PRIVATE_KEY1")
-    bob_private_key = os.getenv("REMOTE_TEST_PRIVATE_KEY2")
-
-    instrs = "You must set it. It must hold Mumbai MATIC."
-    assert alice_private_key, f"Need envvar REMOTE_TEST_PRIVATE_KEY1. {instrs}"
-    assert bob_private_key, f"Need envvar REMOTE_TEST_PRIVATE_KEY2. {instrs}"
-
-    # wallets
-    n_confirm, timeout = config["BLOCK_CONFIRMATIONS"], config["TRANSACTION_TIMEOUT"]
-    alice_wallet = Wallet(web3, alice_private_key, n_confirm, timeout)
-    bob_wallet = Wallet(web3, bob_private_key, n_confirm, timeout)
-
-    print(f"alice_wallet.address = '{alice_wallet.address}'")
-    print(f"bob_wallet.address = '{bob_wallet.address}'")
-
-    return (alice_wallet, bob_wallet)
-
-
 def _remote_config_mumbai(tmp_path):
     config = {
         "RPC_URL": "https://rpc-mumbai.maticvigil.com",
@@ -157,24 +98,6 @@ def _remote_config_mumbai(tmp_path):
 
     # -ensure config is truly remote
     assert "mumbai" in config["RPC_URL"]
-    assert "oceanprotocol.com" in config["METADATA_CACHE_URI"]
-    assert "oceanprotocol.com" in config["PROVIDER_URL"]
-
-    return config
-
-
-def _remote_config_polygon(tmp_path):
-    config = {
-        "RPC_URL": "https://polygon-rpc.com",
-        "BLOCK_CONFIRMATIONS": 0,
-        "TRANSACTION_TIMEOUT": 60,
-        "METADATA_CACHE_URI": "https://v4.aquarius.oceanprotocol.com",
-        "PROVIDER_URL": "https://v4.provider.polygon.oceanprotocol.com",
-        "DOWNLOADS_PATH": "consume-downloads",
-    }
-
-    # -ensure config is truly remote
-    assert "polygon" in config["RPC_URL"]
     assert "oceanprotocol.com" in config["METADATA_CACHE_URI"]
     assert "oceanprotocol.com" in config["PROVIDER_URL"]
